feat(Homebrew): AND-1319 Change countryList to an extension function
